+++ fiddles/angular2-cli/fiddle-0017-StockCarousel/src/app/line-chart/intraday.service.ts
@@ -12,25 +12,25 @@ export class ComplaintsWithPercent {
 }
 
 let complaintsData: Complaints[] = [{
-  complaint: "Pizza is cold",
+  complaint: 'Pizza is cold',
   count: 780
 }, {
-  complaint: "Inadequate cheese quantity",
+  complaint: 'Inadequate cheese quantity',
   count: 120
 }, {
-  complaint: "Not baked properly",
+  complaint: 'Not baked properly',
   count: 52
 }, {
-  complaint: "Delayed delivery",
+  complaint: 'Delayed delivery',
   count: 1123
 }, {
-  complaint: "Damaged delivery",
+  complaint: 'Damaged delivery',
   count: 321
 }, {
-  complaint: "Incorrect billing",
+  complaint: 'Incorrect billing',
   count: 89
 }, {
-  complaint: "Wrong size delivered",
+  complaint: 'Wrong size delivered',
   count: 222
 }];