+++ fiddles/angular2-seeder/fiddle-0005-StockCarousel/src/client/app/quote/quote.component.ts
@@ -1,0 +1,15 @@
+import {
+  Component
+} from '@angular/core';
+
+import { BaseComponent } from '../base.component';
+
+@Component({
+  selector: 'app-quote',
+  moduleId: module.id,
+  templateUrl: './quote.component.html',
+  styleUrls: ['./quote.component.css']
+})
+export class QuoteComponent extends BaseComponent {
+
+}